+++ gradle-ext/main/kotlin/com/javiersc/gradle/extensions/Logging.kt
@@ -1,29 +1,31 @@
 package com.javiersc.gradle.extensions
 
 import com.javiersc.kotlin.stdlib.AnsiColor
+import com.javiersc.kotlin.stdlib.AnsiColor.Foreground
+import com.javiersc.kotlin.stdlib.AnsiColor.Reset
 import com.javiersc.kotlin.stdlib.ansiColor
 import org.gradle.api.logging.Logger
 
-fun Logger.lifecycleColored(color: AnsiColor = AnsiColor.Reset, message: () -> String) {
+public fun Logger.lifecycleColored(color: AnsiColor = Reset, message: () -> String) {
     lifecycle(message().ansiColor(color))
 }
 
-fun Logger.traceColored(color: AnsiColor = AnsiColor.Foreground.Purple, message: () -> String) {
+public fun Logger.traceColored(color: AnsiColor = Foreground.Purple, message: () -> String) {
     trace(message().ansiColor(color))
 }
 
-fun Logger.debugColored(color: AnsiColor = AnsiColor.Foreground.Cyan, message: () -> String) {
+public fun Logger.debugColored(color: AnsiColor = Foreground.Cyan, message: () -> String) {
     debug(message().ansiColor(color))
 }
 
-fun Logger.infoColored(color: AnsiColor = AnsiColor.Foreground.Blue, message: () -> String) {
+public fun Logger.infoColored(color: AnsiColor = Foreground.Blue, message: () -> String) {
     info(message().ansiColor(color))
 }
 
-fun Logger.warnColored(color: AnsiColor = AnsiColor.Foreground.Yellow, message: () -> String) {
+public fun Logger.warnColored(color: AnsiColor = Foreground.Yellow, message: () -> String) {
     warn(message().ansiColor(color))
 }
 
-fun Logger.errorColored(message: String, color: AnsiColor = AnsiColor.Foreground.Red) {
+public fun Logger.errorColored(message: String, color: AnsiColor = Foreground.Red) {
     warn(message.ansiColor(color))
 }