+++ gradle-ext/main/kotlin/com/javiersc/gradle/extensions/Project.kt
@@ -1,0 +1,6 @@
+package com.javiersc.gradle.extensions
+
+import org.gradle.api.Project
+
+public val Project.isRootProject: Boolean
+    get() = parent == null